[enhancement] Docker: Add docker support w Dockerfile & docker-compose (#98)
* chore: add .dockerignore, Dockerfile, and docker-compose.yaml files for Docker containerization

The .dockerignore file is added to specify files and directories that should be excluded from the Docker build context.

The Dockerfile is added to define the steps to build the Docker image. It starts from the node:14 base image, sets the working directory, copies package.json and yarn.lock, installs dependencies, copies the rest of the files, builds the project, and exposes port 3000. The CMD instruction specifies the command to run when the container starts.

The docker-compose.yaml file is added to define the services for the Docker containers. It includes a service for the geohub application, which builds the image using the current directory as the build context, sets environment variables from the .env file, mounts the .env file as a volume, and maps port 3000. It also includes a service for a MongoDB database, which uses the official mongo image, restarts always, sets environment variables from the .env file, and mounts a volume for the database data.

* docs(README.md): add instructions for running the app using Docker and Docker Compose

Added instructions for running the app using Docker and Docker Compose. The instructions include cloning the repository, obtaining a Google Maps API key, generating secrets for NextAuth and Cryptr, creating a `.env` file with the necessary environment variables, and running the app using the `sudo docker compose up` command.

* chore(.gitignore): add 'database/' to ignore list to exclude Docker volumes from version control

* feat(Docker): Add github actions build

* feat(Docker action): Bump versions and fix issues

* fix(Docker action): Typo

* typo again

* feat(Docker): Use built image instead of building

* fix(Docker): Typo when copy pasting